Split to components and hooks
